% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Print FizzBuzz for a given number}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{any number greater than 0}
}
\value{
numbers 0 to n, with "Fizz", "Buzz", or FizzBuzz where appropriate
}
\description{
Prints "Fizz" for every factor of 3, prints "Buzz" for every factor of 5 and
"FizzBuzz" for every factor of 3 and 5.
}
\examples{
FizzBuzz(20)
FizzBuzz(16)
}
