% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Flagstaff Weather}
\format{
A data frame with 1719 observations and 5 columns.
\describe{
\item{DATE}{The days that the data was gathered on}
\item{PRCP}{The Percipitation on that day}
\item{SNOW}{If there was snow on that day and how much}
\item{TMAX}{The maximum temperature for that day}
\item{TMIN}{The minimum temperature for that day}
}
}
\source{
\url{https://github.com/BuscagliaR/STA_444_v2/tree/main/data-raw}
}
\usage{
Flagstaff_Weather
}
\description{
Weather data collected from the Pulliam Airport for the duration of 2015,
2016, 2017, 2018, and 2019.
}
\keyword{datasets}
